requirements.txt.txt
requests
tweepy
yfinance
beautifulsoup4
Pillow
python-dotenv
